refactor(build): Break apart `Build` with a before step, `Prepare`
This commit breaks up the larger `Build` method of the only-implementing
interface for `builder` with a new method, `Prepare`.  This is intended
to break up the preparation of artifacts with the build.  Additionally,
it allows relevant artifacts from a build to be fetched, thus the
application/project structure to be fully realized.  It is important
to do this in fact before building the root filesystem such that we can
determine the full kconfig and whether the rootfs intends to be bundled
with the kernel via embedded initrd functionality.

Signed-off-by: Alexander Jung <alex@unikraft.io>