feat(github-action): Build rootfs based on target architecture
This commit introduces the ability to build a root filesystem based
on the supplied target's architecture.  This makes use of the new
option-method `WithArchitecture` which handles the hint.  Further,
since the root filesystem can be embedded as part of the initramfs,
the construction of the root filesystem is migrated to before the
build and not before the packaging.

Signed-off-by: Alexander Jung <alex@unikraft.io>