fix(paraprogress): Do not quit on receiving 100% progress
This may pre-emptively shutdown the process.  Instead, let the
process exit "naturally".  If the process is reporting via the
onProgress-callback 100%, then this a mistake of the use of the
callback itself, and the user will simply have to wait beyond
the 100% until the process exits.  There is now also a
`WithTimeout` option so we can have both graceful and non-
graceful exits with this package.

Signed-off-by: Alexander Jung <alex@unikraft.io>