fix(paraprogress): Fix restoring the IOStreams (stdout) output
Previously, the context was deeply copied in the attempt to create
per-process working versions which could be passed to individual
processes with slight modifications (specifically to contextual
systems that worked with stdout).  This did not always work which
resulted in very occassional malformed or not-fully-copied contexts.
Instead, simply perform a shallow copy such that full references
can be made per-process and defer a restoration of the iostreams
and stdout after the paraprogress exits.  This is much more
reliable.

Signed-off-by: Alexander Jung <alex@unikraft.io>