fix(buildenvs): Ensure build environment's target a platform
To ensure consistency across different hosts which may vary in
platform (OS and architecture), ensure that there is a variable that
is set to a known default (`linux/x86_64`).  Whilst in the future we
may support multi-platform images, for now, this allows building
such environments on different platforms and ensuring repeatable
outcome of the image itself.

Signed-off-by: Alexander Jung <alex@unikraft.io>