refactor(pull)!: Distinguish between `-o` and `-w`
This commit reworks the flags `-o|--output` and `-w|--workdir` which
can be used to signify two different distinct requirements when
pulling a package.  The `-w|--workdir` flag is used to identify a
location where a `Kraftfile` is and therefore potentially where
package definitions are.  In most cases, a user will simultaneously
wish for the output of pulling packages to this working directory,
since it represents the project directory.  Previously, the
`-o|--output` flag was available but unused.  Now when unset, the
output directory and therefore location of package artifacts is set
to the working directory.  A user can now customize the output
location by changing this flag value, and therefore saving package
artifacts outside of the project directory.

Additionally, since pulling does not necessarily need a working
directory, since a named package can be used as a positional
argument, the `-o|--output` flag better reflects the intent of the
user.

Signed-off-by: Alexander Jung <alex@unikraft.io>