feat(deploy): Introduce deployer which handles full Unikraft builds
This deployer handles a context where the user wishes to construct the
kernel, and as a result the `unikraft` attribute of the `Kraftfile`
is present.  Re-use the build method and options from the `kraft build`
sub command as well as the runtime packager which will understand
a context that has a built unikernel.

Signed-off-by: Alexander Jung <alex@unikraft.io>