build: Improve Make target for tools
This commit reworks the non-phony makefile targets for binaries
which are listed under the `TOOLS` variable and the tools/
directory.  The rework largely follows how `BIN` binaries of the
project are built which: a). include LDFLAG `-X` link-time
string annotations which dictate the instantatious version
(which will fix an issue where "No version provided" is included
in OCI images built via the GitHub Action) and b). to be able
to have a switch to include debug symbols.

Signed-off-by: Alexander Jung <alex@unikraft.io>