fix: Enable network user mode in QEMU
To enable the use of `hostfwd`, the `slirp` module is needed.
This can no longer be found as a static dependency in QEMU 7.2.0 and
above.

As such, we downgrade QEMU from `7.2.4` to `7.1.0`, enable `slirp` via
git, and pin the OS version to the ones in base.

QEMU `7.1.0` is still a valid option and we can use it until we find it
too out of date. For example, Ubuntu 22.04 LTS uses QEMU `6.1.0`.

Signed-off-by: Alexander Jung <alex@unikraft.io>
Signed-off-by: Cezar Craciunoiu <cezar@unikraft.io>