refactor: Clean up `kraft pkg info` by using native `Catalog`
This commit effectively reverses #536 in favour of using the
existing package manager's `Catalog` method.  Since packages
already expose a `Columns` method, this information can be used
to "show" information about a specific named package.
Simultaneously, the `Catalog` method and the information returned
by the implementing `pack.Package` works for all current package
managers.

Secondly, wrap the invocations of `Catalog` with `tui.processtree`
which is useful when used with the `-u|--update` flag for very
long wait times and to help indicate to the user the longevity
of the invocation.

Finally, utilize the newly introduced `PrintPackages` method
which was previously ripped from `kraft pkg ls`.

Signed-off-by: Alexander Jung <alex@unikraft.io>