fix(app): Parse the resulting target's new `.config*`
Following a successful call to `Configure()`, the resulting artifact
is a `.config*` file for the given target which contains the exact
KConfig options which are used for this particular target as understood
by Unikraft's build system.  Add an additional post-processing sequence
which reads this into the target's KConfig map.

This fix enables correct propagation of KConfig options as metadata
across the codebase, fixing several places where the data was not
available, including packaging in the GitHub Action.

Signed-off-by: Alexander Jung <alex@unikraft.io>