fix(initrd): Set the root as positional to the current working directory
At the present, CPIO archives are built using an absolute path.  For
debugging purposes, this makes it difficult to extract locally.  Instead,
these archives should be built as relative paths (i.e. ./ instead of /)
to the current working directory.

Simultaneously, setting the path as relative does not affect whether the
archive is mounted at root, since it is now relative to root.  This also
makes it possible to extract archives at subdirectories (relative to that
directory).

GitHub-Fixes: #1037
Signed-off-by: Alexander Jung <alex@unikraft.io>