refactor(build)!: Add `Target` and `TargetName` as `BuildOptions`
This commit reworks the `BuildOptions` struct such that you can pass a
string `TargetName` but also a previously populated `Target`, which
subsequently prevents the dynamic selection process to occur.  At the
same time, because the Target can be provided, the interface method for
the `builder` interface is updated since the previously passed
positional argument `targ` is now part of `opts`.

Signed-off-by: Alexander Jung <alex@unikraft.io>