fix(tui): Correctly handle new `ctx` with custom log and iostreams out
This commit fixes an issue where the "stdout" of both the general
logger and the iostreams was incorrectly set for all "subprocesses"
of `paraprogress` and `processtree`.  Since the `context.Context`
object was not unique per process, all subprocesses would output to
the last process (becauase of the iteration process which attempted
to set the output writer).

To make it so that each context is unique, perform a deepcopy of
only the log and iostreams context objects and modify appropriately
with new values.  The rest of the context is re-used as-is, allowing
subprocesses to continue to access values as necessary.

Signed-off-by: Alexander Jung <alex@unikraft.io>