feat: Introduce `Unpack` interface method to `pack.Package`
This commit introduces a new method to all package implementations,
`Unpack`, which allows for the separation of pulling and unpackaging
into two distinct steps.  Note that for now, this is only consumed at
the application-level with the `runner_package.go` (`kraft run`) and
therefore OCI packages and not fully implemented or Manifest-type
packages.

Providing such a distinction speeds up the instantiation of pre-pulled
packages representing applications.

Signed-off-by: Alexander Jung <alex@unikraft.io>