refactor!: Distinguish between "local" and "remote" queries
This commit deprecates the package manager's query option for
`WithUpdate` and splits it into two distinct options: `WithLocal` and
`WithRemote` where local queries are by default `true` and `WithLocal`
can be used to set local queries to false and prevent local lookups.

Accordingly, the `Catalog` method for OCI and Manifest package managers
are updated to reflect this.

Signed-off-by: Alexander Jung <alex@unikraft.io>