fix(run): Remove erronous `else` statement
This commit removes a left-over `else` statement that inadvertently
prevent the `kraft run` command from executing the sibling method which
a previous refactor aimed to incorporate.  Instead, use of the old
machine controller was used.  Whilst this does start the machine, does
not guarantee proper execution in the context of the CLI where the
sibling `Start` method does (e.g. proper tailing of logs).

Signed-off-by: Alexander Jung <alex@unikraft.io>