feat(cache): Introduce `RemoteIndex`
`RemoteIndex` is a wrapper for `v1.RemoteIndex` which caches requests
in-memory for previously requested indexes.  This aids in reducing the
number of lookup calls to the same index.  Since `v1.WithPlatform` is not
respected, a valid lookup will have had any additional options, such as
`WithTransport` and `WithAuth`, fully satisfied.

Signed-off-by: Alexander Jung <alex@unikraft.io>