feat(cache): Introduce `RemoteImage`
`RemoteImage` is a wrapper for `v1.Image` which caches requests
in-memory for previously requested image.  This aids in reducing
the number of lookup calls to the same images.  Since
`v1.WithPlatform` is not respected, a valid lookup will have had
any additional options, such as `WithTransport` and `WithAuth`,
fully satisfied.

Signed-off-by: Alexander Jung <alex@unikraft.io>