refactor(oci): Improve `NewPackageFromOCIManifestDigest`
This commit reworks parts of `NewPackageFromOCIManifestDigest` to use
parallelizm when looking up indexes containing multiple manifest images
which can be handled concurrently as well correctly seeding unknown
manifests that could not be determined locally.  This generally improves
the performance and reliability of this method when passing in
references.

Additionally, the recently introduced caching of v1(Index)Image is used
accordingly.

Signed-off-by: Alexander Jung <alex@unikraft.io>