feat(oci): Handle partial (non-distributable) OCI layers
This commit reworks internal mechanisms in the OCI package such that it
is capable of handling packages which simply _reference_ a layer without
having to pull or upload it.  This is both an optimization step in terms
of preventing users who are simply re-packaging unikernels with new
rootfilesystem layers to download and re-upload the same kernel layer as
well as redundant cache coping mechanism such that remote references to
kernel layers are kept up-to-date and re-packaging ensures the latest
reference.

Numerous changes were necessary across the OCI package manager to make
this possible, including:

- keeping an in-memory copy of previously instantiated `ociPackage`
  structure;
- preventing error-outs from not found layers (which are now logged as
  non-distributable);
- better handling of saving and updating of the `manifest` and `index`
  internal structures if and when changes to the manifest and index
  exactly occur; and,
- "re-tagging" non-local, "non-distributable" OCI layers (in compliance
  with the OCI Distribution Specification) during the `Push` procedure.

Signed-off-by: Alexander Jung <alex@unikraft.io>