fix(handler): Always remove the manifest
This commit fixes an issue where the manifest is not deleted if any of
its dependencies have failed to be removed.

Note: this may lead to lingering digests, but during testing I could not
detect this.  Instead, errors were generally to do with missing
dependencies as opposed to inability to remove them.

Signed-off-by: Alexander Jung <alex@unikraft.io>