refactor(pkg): Rework the runtime packager to support non-pulls
This commit reworks the subcommand for `kraft pkg` when a `Kraftfile`
specifies a `runtime` element with a new flag `--no-pull`.  Ultimately
this stems from the rework done in the package manager and specifically
the OCI package manager and its ability to reference non-distributable
OCI image layers.

To accommodate this change, the packager is updated to use the `Save`,
`IsPulled` and `Remote`/`Local` interface methods from the
`pack.Package` and performs an appropriate search before re-packaging
the filesystem with a remote package reference.

Signed-off-by: Alexander Jung <alex@unikraft.io>