fix(oci): Better handling of partial indexes
This commit migrates the functionality of
`NewPackageFromOCIManifestDigest` which handles remote access of an
index and its manifest into `newIndexAndManifestFromRemoteDigest`.  This
method is then used during a re-attempt in accessing an index and its
manifest if a local reference cannot be instantiated.  This increases
the overal resilience of the exposed method.

Signed-off-by: Alexander Jung <alex@unikraft.io>