feat(deploy): Mechanism to switch between detected deployer candidates
In order to improve usability, it is sometimes possible that mulitple
deployer contexts are valid for a given current working directory and
set of input arguments.

To combat this, instead of error-ing out with a vague message, allow the
user to switch between the contexts when prompting/interactivity is
enabled.  This allows the user to continue.

Signed-off-by: Alexander Jung <alex@unikraft.io>