feat(initrd): Dynamically instantiate BuildKit in container
This commit introduces the ability to dynamically start BuildKit using
the testcontainer framework[0] which provides an abstraction over,
at least at the time of commit, Docker, Docker Desktop and Podman.

Its ability to interface with the host and start a container based on
these providers may able to provide to the vast set of users an
ease-of-use improvement which would otherwise have resulted in the
Dockerfile initrd provider to spit a warning, requiring additional
user interaction.  This has caused many users needless hassle and
so this commit aims to solve this by dynamically starting BuildKit
and connecting to it.

Since the container may still linger from quick successions of `kraft`
invocations, dynamically allocate a port for BuildKit between 4000-
5000.

[0]: https://testcontainers.com/

Signed-off-by: Alexander Jung <alex@unikraft.io>