fix(run): Improve filter and selection process of potential packages
This commit partially reworks the filtering of possible runtime packages
from the given input for the Kraftfile-with-runtime context.  In this
rework;

a). the initial query to the `Catalog` is wrapped in a process tree
    which hides the search logs from the catalog, reduding the visible
    output in "fancy-log" mode;
b). uses the provided CLI flags for platform and architecture later in
    the query process such that the option of selecting a package with
    alternative platform/architecture is possible.

Ultimately this commit improves the experience of running and selecting
a possible runtime following a `kraft run` invocation.

Signed-off-by: Alexander Jung <alex@unikraft.io>