fix(oci): Pass query to internal `update` method
Ultimately passing the query to the internal `update` method means
that we can pass the query to `processV1IndexManifest` which contains
a filtering process based on the platform, architecture and provided
set of KConfig options.  This propagates appropriately to the
`Catalog` method when a remote set of indexes is requested, ultimately
providing a more accurate set of resulting packages.

Signed-off-by: Alexander Jung <alex@unikraft.io>