fix(action): Elevate group permissions if unknown UID:GID are used
This moves the check outside the container and makes sure that files are
always accesible. To do this, group permissions are elevated in the
working directories when unknown UID:GID are used. This works because
groups are hardcoded in the container for the 1000, 1001, 1002 GIDs.

To disable this if permissions are to be left untouched set 'grouprwx'
to 'false'.

Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>