fix(qemu): Show first line after reciving "Booting from ROM"
Return the channel if the phrase from QEMU has been received.
Since the line has now been consumed, the next won't be if
we simply `continue`, because this will be consumed to be
checked.  Without this, the first line from the QEMU log
will be lost.  For applications which output a single line,
this would otherwise be lost.

Signed-off-by: Alexander Jung <alex@unikraft.io>