fix(run): Prevent possible nil-pointers by always using `Cut`
When parsing environmental variables in `kraft run`, which are
often presented in circumstances with strings with an equals
delimeter, the use of `SplitN` was not always checked for empty
values.  To prevent possible nil-pointers and to keep code
consistent, replace all areas where this type of procedure
occurs with `strings.Cut` instead.

Signed-off-by: Alexander Jung <alex@unikraft.io>