fix(oci): Don't delete dependencies of index
When we package an application it can be the case that we have
a dependency on another package (e.g. a kernel). In these cases,
the dependency is not actually packed together with our application,
instead we just keep a reference to it. The problem arises when we
attempt to overwrite such a package. Because all dependencies
were previously deleted, this meant that both of the underlying packages
would be removed, but only the application would get repackaged,
resulting in an app that cannot be run due to missing layers.

Setting this flag to false and not removing dependencies solves this
issue, but causes the cache to grow indefinitely, which is less than
ideal.

Signed-off-by: Luca Seritan <luca.seritan@gmail.com>