fix(webinstall): Non-package install dependencies in manual step
This commit fixes a misplacement of the depdendency installation
methods which should only occur on a manual installation (i.e.
not via package manager).  Since package manager steps include
in themselves a depedendency question this a). would mean that
in package manager mode, the question appears twice and b).
in manual mode, the option is not available.

Signed-off-by: Alexander Jung <alex@unikraft.io>