fix(compose): Do not associate networks with service groups
This refactor removes the association of "compose networks" with
KraftCloud service groups.  This association is in fact incorrect
as the service groups are intended to be used for a single
application as opposed to a group of networks; to be used for
autoscaling and load-balancing for said single application.

As a result of this refactor, introduce a new non-attribute
`Services` to the compose create routine such that we can still
propagate ports correctly.

Additionally, adjust the `kraft cloud compose down` CLI subcommand
as it does not need to remove associated service groups anymore.

Signed-off-by: Alexander Jung <alex@unikraft.io>