feat(cmdfactory): Recognize `types.ShellWords` in the command-line
This type is specifically used within compose and Docker's internal
tooling such that it can be recognised a). as its own type and b).
such that it is nil-able without it being empty.  Unlike a
`[]string` type, `nil` also signals that the slice is of size 0.
In this case, we want to also detect that it is not set at the
command-line.  This will allow commands to use the type and
know whether it is nil, it has been set and is specifically empty
or is non-empty.

Signed-off-by: Alexander Jung <alex@unikraft.io>