feat(create): Introduce `--entrypoint` flag
This is used to overwrite the image's arguments.  Similar to
Docker and `Dockerfile`'s, the image has a supplied default command
which can be referenced and used.  With the use of
`types.ShellCommand` we can also detect whether an empty set of
command should be used.

The new default functionality is to accept the set of positional
arguments as appends to the images arguments.  The user can
overwrite this by setting a value to `--entrypoint`.  Supplying
an empty `--entrypoint` zeros-out that image value.

Fetch the metadata of the image beforehand to ensure we correctly
set everything too.

Signed-off-by: Alexander Jung <alex@unikraft.io>