fix(build): Better handling of compose builds
This fix adds more logic to the `kraft cloud compose build`
subcommand to better stabilize it.  Specifically, it tries
to handle the case where the compose spec does not provide
a "build" which translates to it referencing an image.
However, these images may be valid OCI images which are not
unikernels.  A look up is performed to determine whether
there is a KraftCloud version (both from the user and
officially) of this image equivilant.  If neither, then
a build (or at least a "repackage") _can_ be performed
by this toolchain in order to convert the traditional
OCI image into a valid unikernel OCI.

Signed-off-by: Alexander Jung <alex@unikraft.io>