feat(deploy): Introduce "rootfs" deployer
This commit introduces a new deployer which is able to deploy a
context where no working directory with no Kraftfile is present
and either a `--rootfs` flag is supplied or a detected
`Dockerfile` in the current working directory is detected.

In this scenario we can dynamically create a new project and
supply related information such that we are able to from this
basic scenario create a new Unikraft Cloud deployment.

Signed-off-by: Alexander Jung <alex@unikraft.io>