fix(deploy): Prevent deployers with empty `String()` from prompting
Simultaneously, remove the "rootfs" deployer from being promptable.
In a circumstance where the rootfs is promptable, there are other
potential deployers.  These deployers should in fact take priority
and the rootfs deployer should not be made available to the user
as it relies heavily on CLI flags and arguments which are otherwise
provided by other deployers.

Signed-off-by: Alexander Jung <alex@unikraft.io>