fix(instance): Update instance flag for volumes
Signed-off-by: Todor Todorov <todor@todorov.uk>

The 'kraft cloud inst create' accepts flag '--volumes' which
could mean that we can pass pultiple volumes as value to the
flag.
In fact, the flag can accept only one volume, but can be used
multiple times.
Also, in the example we can see the flag being used is '--volume'.