fix(cloud): Correctly set default enum attributes
This commit fixes an issue whereby the `Create` command which can be
used by other inner subcommands does not set the default value of
the enums for restart and rollout policies.  In a command-line-context
these are set automatically, but not when invoked indirectly
internally from another subcommand (e.g. from compose subcommands).

Signed-off-by: Alexander Jung <alex@unikraft.io>