fix(kraft): Handle better error messages and prevent preemptive checks
This commit essentially reverts the changes made in d5990f336 and
b34585f24 as preemptive checks for architecture and platform should
not be performed at this stage.  These commits address a real issue
which is undescriptive error messages (which is tackled in this new
commit).  The reason for not wishing to preemptively check for the
architecture and platform is that these can be dynamically
determined based on the `packager` and `builder` contexts which
involve retrieving a remote runtime and can contain values which
are not directly managed by KraftKit.

Signed-off-by: Alexander Jung <alex@unikraft.io>