fix(paraprogress): avoid race condition
It seems possible to trigger a weird race condition in which the status
is not marked as success/failed, and so the progress spinner never exits.
This previously wasn't a problem, since it seems like we would exit when
the progress reached 1.0. However, a recent change modifies that.

My *hunch* is that batching all the cmds in start is causing this - these
are run in parallel, so could be racy. It seems possible that:
- The process is run
- The process finished and we send StatusMsg = success
- We send StatusMsg = running, resetting the state back to running
This means that a process can end up in a running state forever.

This is definitely a *possible* race condition (that's worth fixing), I'm
just not 100% sure it's exactly the one Artūrs was seeing.

Signed-off-by: Justin Chadwell <justin@unikraft.com>